xml serialization fine tuned.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2632 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/TableTypes/Document.cs
@@ -114,9 +114,9 @@ namespace OpenDentBusiness{
 
 		///<summary>Only allowed 0,90,180, and 270.</summary>
 		[DataField("DegreesRotated")]
-		private short degreesRotated;
+		private int degreesRotated;
 		/// <summary>Only allowed 0,90,180, and 270.</summary>
-		public short DegreesRotated {
+		public int DegreesRotated {
 			get { return degreesRotated; }
 			set { degreesRotated = value; MarkDirty(); degreesRotatedChanged = true; }
 		}
